Add reusable bounty markdown renderer
