+++ frontend/package.json
@@ -28,6 +28,7 @@
     "react-hook-form": "^7.71.1",
     "react-markdown": "^10.1.0",
     "react-simple-typewriter": "^5.0.1",
+    "remark-gfm": "^4.0.1",
     "sonner": "^2.0.7",
     "tailwind-merge": "^2.6.0",
     "xumm": "^1.8.0",
