+++ frontend/src/app/[locale]/bounties/create/page.tsx
@@ -4,8 +4,7 @@ import React, { useState } from "react";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import * as z from "zod";
-import ReactMarkdown from "react-markdown";
-import { CodeBlock } from "@/components/markdown/CodeBlock";
+import { BountyDescription } from "@/features/bounties/components/BountyDescription";
 import {
   Edit3,
   Eye,
@@ -180,17 +179,7 @@ export default function CreateBountyPage() {
                   <h2 className="text-4xl font-black italic uppercase tracking-tighter mb-8">
                     {formData.title || "Untitled Mission"}
                   </h2>
-                  <div className="prose prose-invert max-w-none prose-violet">
-                    <ReactMarkdown
-                      components={{
-                        pre: ({ children, className }) => (
-                          <CodeBlock className={className}>{children}</CodeBlock>
-                        ),
-                      }}
-                    >
-                      {formData.description || "_Intel pending..._"}
-                    </ReactMarkdown>
-                  </div>
+                  <BountyDescription markdown={formData.description} />
                 </motion.div>
               )}
             </AnimatePresence>
