+++ frontend/src/app/[locale]/layout.tsx
@@ -1,13 +1,9 @@
 import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import "../globals.css";
 import { getMessages } from 'next-intl/server';
 import { NextIntlClientProvider } from 'next-intl';
 import { locales, defaultLocale, type Locale } from '@/i18n';
 import { ErrorBoundary } from '@/components/ErrorBoundary';
-import { ThemeProvider } from "next-themes";
-
-const inter = Inter({ subsets: ["latin"] });
+import { getDirection } from "@/lib/i18n/utils";
 
 export const metadata: Metadata = {
   title: "Stellar Guilds",
@@ -29,27 +25,20 @@ export default async function RootLayout({
     locale = defaultLocale;
   }
 
-  // Get the messages for the current locale
   const messages = await getMessages();
+  const direction = getDirection(locale as Locale);
 
   return (
-    <html lang={locale} suppressHydrationWarning>
-      <body className={inter.className}>
-        <ThemeProvider
-          attribute="class"
-          defaultTheme="dark"
-          enableSystem={false}
+    <NextIntlClientProvider locale={locale} messages={messages}>
+      <ErrorBoundary>
+        <div
+          dir={direction}
+          className="min-h-screen flex flex-col bg-white dark:bg-stellar-navy text-gray-900 dark:text-stellar-white font-sans transition-colors duration-300"
         >
-          <NextIntlClientProvider locale={locale} messages={messages}>
-            <ErrorBoundary>
-              <div className="min-h-screen flex flex-col bg-white dark:bg-stellar-navy text-gray-900 dark:text-stellar-white font-sans transition-colors duration-300">
-                {children}
-              </div>
-            </ErrorBoundary>
-          </NextIntlClientProvider>
-        </ThemeProvider>
-      </body>
-    </html>
+          {children}
+        </div>
+      </ErrorBoundary>
+    </NextIntlClientProvider>
   );
 }
